Updating content to avoid accessibility fail
Collaborating with Julian Panico Team B's content designer we have update the page structure to improve the page's accessibility in terms of structure and content.

The hint text is too long and isn't actually proper hint text. Hint text i designed to assist users in how to fill in the form correctly. This content is simply additional helpful content and as hint text adds noise to user of assistive tenchology.
So I have changed the page structure to the recommended accessibility solution of:
h1
p
fieldset
visually hidden legend repeats h1
radio buttons

This is also a boolean Yes/No pattern style question page, so we added that in to clarify for users and updated the error message to match the GOV.UK prescribed Yes/No radio pattern.